requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==42.0.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyMySQL==1.1.0
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
sse-starlette==2.2.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
